Minor bug fixes

+++ src/main/java/entidad/huespedIVA.java
@@ -3,10 +3,10 @@ package entidad;
 public class huespedIVA {
     private String ocupacion;
     private String posicionIVA;
-    private int cuit;
+    private long cuit;
     private boolean responsableDePago;
 
-    public huespedIVA(String ocupacion, String posicionIVA, int cuit, boolean responsableDePago) {
+    public huespedIVA(String ocupacion, String posicionIVA, long cuit, boolean responsableDePago) {
         this.ocupacion = ocupacion;
         this.posicionIVA = posicionIVA;
         this.cuit = cuit;
@@ -19,8 +19,8 @@ public class huespedIVA {
     public String getPosicionIVA() { return posicionIVA; }
     public void setPosicionIVA(String posicionIVA) { this.posicionIVA = posicionIVA; }
 
-    public int getCuit() { return cuit; }
-    public void setCuit(int cuit) { this.cuit = cuit; }
+    public long getCuit() { return cuit; }
+    public void setCuit(long cuit) { this.cuit = cuit; }
 
     public boolean isResponsableDePago() { return responsableDePago; }
     public void setResponsableDePago(boolean responsableDePago) { this.responsableDePago = responsableDePago; }
